Add boilerplate
Added React
Added kube css framework
Update webpack for less css
