+++ .babelrc
@@ -1,4 +1,4 @@
 {
-  "presets": ["env"],
+  "presets": ["env", "react"],
   "plugins": ["transform-object-rest-spread"]
 }
