+++ src/components/App.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+const App = () => <h1>Hello World</h1>;
+
+export default App;
