+++ src/index.html
@@ -3,12 +3,11 @@
 
 <head>
     <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
 </head>
 
 <body>
-    <div id="root">
-
-    </div>
+    <div id="root"></div>
 
     <script src="/bundle.js"></script>
 </body>
