+++ src/style.less
@@ -1,0 +1,1 @@
+@import './vendor/kube.css';
